Can now search between 2 locations using the to and from search box.
Both Search boxes work as expected now.
Camera zooming to display both markers.

To be included in later commits:
Routing between the 2 locations.

Considering to fix:
onResuming and onPause screws up the program's state because it only
resumes the current location. (you can try by rotating the device and it
will reinitialize as the current location, which is not intended as it
loses all the user's progress.)

Might want to include a smarter padding system for the zooming between 2
markers that scales accordingly with the distance between 2 markers.