Solved to resuming issue but unable to solve screen rotation issue
If screen rotates, it recreates the whole activity losing all the stored
data. I realised that the edittext still stores the data but I was not
successful in implementing recovery.

Hence, I decided to leave it as it is. At least, now if you close the
app screen and open it again, It still resumes the previous screen now.

I included my resources in a word doc.

Future works:
Once, we finalize how to combine the different components of the
activities, I will adjust the drawLine() code to implement the relevant
waypoints and draw the lines using different colours for different modes
of transport