Travel App Map is the latest
Ready for integration

Look at the Search in the activity_map_test for where to include the
arraylist<Integer> of transportsMode and arraylist<String> of locations.

Look at the drawline() method in mapfragment for the color codes.

Do note that the first element of the array for transport mode since I
only start drawing lines when there are 2 points or more. Hence the 2nd
element determines the color of the first line.

If you want to do a return trip to original location, do include the
original location at the end of the array.